refactor codes 리뷰 반영  - authenticateMember 메서드 명 login으로 수정  - 필요없는 주석 삭제  - ExceptionHandler Global하게 처리하도록 수정
