+++ src/main/java/gift/controller/MemberController.java
@@ -5,8 +5,6 @@ import gift.dto.MemberRequestDto;
 import gift.service.MemberService;
 import org.apache.coyote.BadRequestException;
 import org.apache.tomcat.websocket.AuthenticationException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -17,7 +15,6 @@ import org.springframework.web.bind.annotation.*;
 public class MemberController {
 
     private final MemberService memberService;
-    private static final Logger log = LoggerFactory.getLogger(MemberController.class);
 
     MemberController(MemberService memberService) {
         this.memberService = memberService;
@@ -37,15 +34,7 @@ public class MemberController {
             @Validated @RequestBody MemberRequestDto memberRequestDto
     ) throws AuthenticationException {
         return ResponseEntity.ok(
-                memberService.authenticateMember(memberRequestDto)
+                memberService.login(memberRequestDto)
         );
-    }
-
-    @ExceptionHandler(AuthenticationException.class)
-    public ResponseEntity<String> handleAuthenticationException(AuthenticationException e) {
-        log.trace(e.getMessage());
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(e.getMessage());
     }
 }
